refactor(authentication): Update package dependencies from player to user
