+++ backend/src/main/java/com/project/G1_T3/authentication/controller/RegisterController.java
@@ -1,12 +1,11 @@
 package com.project.G1_T3.authentication.controller;
 
-import com.project.G1_T3.player.service.UserService;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.project.G1_T3.authentication.model.RegisterRequest;
+import com.project.G1_T3.user.service.UserService;
 
 @RestController
 @RequestMapping("/authentication")
