+++ backend/src/main/java/com/project/G1_T3/authentication/service/JwtService.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
-import com.project.G1_T3.player.model.User;
+import com.project.G1_T3.user.model.User;
 
 import java.security.Key;
 import java.util.Date;
