+++ backend/src/main/java/com/project/G1_T3/config/SecurityConfig.java
@@ -20,7 +20,7 @@ import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
 import com.project.G1_T3.authentication.filter.JwtAuthenticationFilter;
-import com.project.G1_T3.player.service.CustomUserDetailsService;
+import com.project.G1_T3.user.service.CustomUserDetailsService;
 
 import java.util.Arrays;
