+++ backend/src/main/java/com/project/G1_T3/user/model/UserRole.java
@@ -1,4 +1,4 @@
-package com.project.G1_T3.player.model;
+package com.project.G1_T3.user.model;
 
 public enum UserRole {
     PLAYER, ADMIN
